+++ plog-api/src/main/java/com/airbnb/plog/Message.java
@@ -1,0 +1,7 @@
+package com.airbnb.plog;
+
+import io.netty.buffer.ByteBufHolder;
+
+public interface Message extends ByteBufHolder {
+    byte[] asBytes();
+}
