+++ plog-api/src/test/groovy/com/airbnb/plog/MessageImplTest.groovy
@@ -1,0 +1,10 @@
+package com.airbnb.plog
+
+import io.netty.buffer.ByteBufAllocator
+
+class MessageImplTest extends GroovyTestCase {
+    void testCreateAndRead() {
+        def content = 'foo'.bytes
+        assert MessageImpl.fromBytes(ByteBufAllocator.DEFAULT, content).asBytes() == content
+    }
+}
