+++ plog-server/src/main/java/com/airbnb/plog/listeners/ByteBufToMessageDecoder.java
@@ -1,0 +1,18 @@
+package com.airbnb.plog.listeners;
+
+import com.airbnb.plog.MessageImpl;
+import io.netty.buffer.ByteBuf;
+import io.netty.channel.ChannelHandler;
+import io.netty.channel.ChannelHandlerContext;
+import io.netty.handler.codec.MessageToMessageDecoder;
+
+import java.util.List;
+
+@ChannelHandler.Sharable
+class ByteBufToMessageDecoder extends MessageToMessageDecoder<ByteBuf> {
+    @Override
+    protected void decode(ChannelHandlerContext ctx, ByteBuf buf, List<Object> out) throws Exception {
+        buf.retain();
+        out.add(new MessageImpl(buf));
+    }
+}
