--- plog-server/src/test/groovy/com/airbnb/plog/MessageTest.groovy
@@ -1,10 +1,0 @@
-package com.airbnb.plog
-
-import io.netty.buffer.ByteBufAllocator
-
-class MessageTest extends GroovyTestCase {
-    void testCreateAndRead() {
-        def content = 'foo'.bytes
-        assert Message.fromBytes(ByteBufAllocator.DEFAULT, content).asBytes() == content
-    }
-}
